refactor(nova): split the NOVA page into handler and view modules
Codex P2 on #496: nova-page.ts had grown past the ~400-line file
guideline. nova-page.ts keeps HTTP/action/CSRF handling (210 lines);
nova-page-view.ts carries the page model, rendering, confirm section,
and formatting (302 lines). Types re-exported for API stability; no
behavior change (151 tests unchanged and green).

Also records the maintainer's evidence-batching decision in the
sequencing SSOT: no per-PR cloud evidence cycles and no release until
the program's end — one final evidence refresh before release.
